Updated old auth usage
